+++ controllers/contactControllers.js
@@ -1,4 +1,4 @@
-const Contact = require("./schemas/contact");
+const Contact = require("../models/schemas/contact");
 
 const listContacts = async (id) => {
   return Contact.find({owner: id});
